database completo

+++ db-server/src/main/java/br/com/iftm/dbserver/api/DatabaseApi.java
@@ -8,10 +8,10 @@ import org.springframework.web.bind.annotation.PostMapping;
 public interface DatabaseApi {
 
     @PostMapping(value = "/deposit/{id}/{ammount}")
-    Account deposit(@PathVariable("id") Integer id, @PathVariable("ammount") Long ammount);
+    Account deposit(@PathVariable("id") Integer id, @PathVariable("ammount") Double ammount);
 
     @PostMapping(value = "/withdraw/{id}/{ammount}")
-    Account withdraw(@PathVariable("id") Integer id, @PathVariable("ammount") Long ammount);
+    Account withdraw(@PathVariable("id") Integer id, @PathVariable("ammount") Double ammount);
 
     @GetMapping(value = "/balance/{id}")
     Account balance(@PathVariable("id") Integer id);
