+++ db-server/src/main/java/br/com/iftm/dbserver/service/DatabaseService.java
@@ -4,9 +4,9 @@ import br.com.iftm.dbserver.model.Account;
 
 public interface DatabaseService {
 
-    Account deposit(Integer id, Long ammount);
+    Account deposit(Integer id, Double ammount);
 
-    Account withdraw(Integer id, Long ammount);
+    Account withdraw(Integer id, Double ammount);
 
     Account balance(Integer id);
 }
